+++ src/Components/AllData.js
@@ -1,0 +1,100 @@
+const allData = [
+    {
+        id:1,  //also referred as sku
+        image: "https://m.media-amazon.com/images/I/717-CcECtuL._AC_UL480_FMwebp_QL65_.jpg",
+        name: "Stylish Cups",
+        category: "Cups",
+        price: "9.99",
+        description:"Lorem ipsum dolor sit amet consectetur adipisicing elit. Obcaecati, enim aliquid laboriosam eius autem odio minima recusandae fugiat totam aspernatur veritatis laudantium illo alias optio et explicabo impedit delectus mollitia.",
+    },
+    {
+        id:2,
+        image: "https://images.unsplash.com/photo-1579446565308-427218a2c60e?w=600&auto=format&fit=crop&q=60&ixlib=rb-4.1.0&ixid=M3wxMjA3fDB8MHxzZWFyY2h8Nnx8Zm9vdHdlYXJ8ZW58MHx8MHx8fDA%3D",
+        name: "Stylish Sneakers",
+        category: "Footwear",
+        price: "109.99",
+        description:"Lorem ipsum dolor sit amet consectetur adipisicing elit. Obcaecati, enim aliquid laboriosam eius autem odio minima recusandae fugiat totam aspernatur veritatis laudantium illo alias optio et explicabo impedit delectus mollitia.",
+    },
+    {
+        id:3,
+        image: "https://images.puma.com/image/upload/f_auto,q_auto,b_rgb:fafafa,w_500,h_500/global/402200/05/sv01/fnd/IND/fmt/png/X-Ray-2-Square-Superior-Comfort-Sneakers",
+        name: "Stylish Sneakers",
+        category: "Footwear",
+        price: "99.99",
+        description:"Lorem ipsum dolor sit amet consectetur adipisicing elit. Obcaecati, enim aliquid laboriosam eius autem odio minima recusandae fugiat totam aspernatur veritatis laudantium illo alias optio et explicabo impedit delectus mollitia.",
+    },
+    {
+        id:4,
+        image: "https://images.puma.com/image/upload/f_auto,q_auto,b_rgb:fafafa,w_500,h_500/global/402185/03/sv01/fnd/IND/fmt/png/Court-Shatter-Men's-Sportstyle-Sneakers",
+        name: "Stylish Sneakers",
+        category: "Footwear",
+        price: "59.99",
+        description:"Lorem ipsum dolor sit amet consectetur adipisicing elit. Obcaecati, enim aliquid laboriosam eius autem odio minima recusandae fugiat totam aspernatur veritatis laudantium illo alias optio et explicabo impedit delectus mollitia.",
+    },
+    {
+        id:5,
+        image: "https://images.unsplash.com/photo-1543512214-318c7553f230?w=600&auto=format&fit=crop&q=60&ixlib=rb-4.1.0&ixid=M3wxMjA3fDB8MHxzZWFyY2h8Mnx8YW1hem9uJTIwJTIwcHJvZHVjdHxlbnwwfHwwfHx8MA%3D%3D",
+        name: "Amazon Product",
+        category: "Speaker",
+        price: "49.99",
+        description:"Lorem ipsum dolor sit amet consectetur adipisicing elit. Obcaecati, enim aliquid laboriosam eius autem odio minima recusandae fugiat totam aspernatur veritatis laudantium illo alias optio et explicabo impedit delectus mollitia.",
+    },
+    {
+        id:6,
+        image: "https://m.media-amazon.com/images/I/61EroGuU16L._AC_UL480_FMwebp_QL65_.jpg",
+        name: "Stylish Cups",
+        category: "Cups",
+        price: "19.99",
+        description:"Lorem ipsum dolor sit amet consectetur adipisicing elit. Obcaecati, enim aliquid laboriosam eius autem odio minima recusandae fugiat totam aspernatur veritatis laudantium illo alias optio et explicabo impedit delectus mollitia.",
+    },
+    {
+        id:7,
+        image: "https://plus.unsplash.com/premium_photo-1679513691474-73102089c117?w=600&auto=format&fit=crop&q=60&ixlib=rb-4.1.0&ixid=M3wxMjA3fDB8MHxzZWFyY2h8NXx8YW1hem9uJTIwJTIwcHJvZHVjdHxlbnwwfHwwfHx8MA%3D%3D",
+        name: "Stylish Headphones",
+        category: "Headphones",
+        price: "49.99",
+        description:"Lorem ipsum dolor sit amet consectetur adipisicing elit. Obcaecati, enim aliquid laboriosam eius autem odio minima recusandae fugiat totam aspernatur veritatis laudantium illo alias optio et explicabo impedit delectus mollitia.",
+    },
+    {
+        id:8,
+        image: "https://images.puma.com/image/upload/f_auto,q_auto,b_rgb:fafafa,w_500,h_500/global/630988/01/mod04/fnd/IND/fmt/png/PUMA-x-NACHO-Men's-Relaxed-Fit-Tee",
+        name: "Stylish Shirts",
+        category: "Shirt",
+        price: "99.99",
+        description:"Lorem ipsum dolor sit amet consectetur adipisicing elit. Obcaecati, enim aliquid laboriosam eius autem odio minima recusandae fugiat totam aspernatur veritatis laudantium illo alias optio et explicabo impedit delectus mollitia.",
+    },
+    {
+        id:9,
+        image: "https://m.media-amazon.com/images/I/71pXblT9X7L._AC_UL480_FMwebp_QL65_.jpg",
+        name: "Stylish Cups",
+        category: "Cups",
+        price: "20.99",
+        description:"Lorem ipsum dolor sit amet consectetur adipisicing elit. Obcaecati, enim aliquid laboriosam eius autem odio minima recusandae fugiat totam aspernatur veritatis laudantium illo alias optio et explicabo impedit delectus mollitia.",
+    },
+    {
+        id:10,
+        image: "https://images.puma.com/image/upload/f_auto,q_auto,b_rgb:fafafa,w_500,h_500/global/777706/02/mod01/fnd/IND/fmt/png/Morocco-2025-Men's-Football-Away-Jersey",
+        name: "Stylish Shirts",
+        category: "Shirt",
+        price: "59.99",
+        description:"Lorem ipsum dolor sit amet consectetur adipisicing elit. Obcaecati, enim aliquid laboriosam eius autem odio minima recusandae fugiat totam aspernatur veritatis laudantium illo alias optio et explicabo impedit delectus mollitia.",
+    },
+    {
+        id:11,
+        image: "https://images.puma.com/image/upload/f_auto,q_auto,b_rgb:fafafa,w_500,h_500/global/402538/02/sv01/fnd/IND/fmt/png/CA-Pro-Classic-Leather-Sneakers",
+        name: "Stylish Sneakers",
+        category: "Footwear",
+        price: "99.99",
+        description:"Lorem ipsum dolor sit amet consectetur adipisicing elit. Obcaecati, enim aliquid laboriosam eius autem odio minima recusandae fugiat totam aspernatur veritatis laudantium illo alias optio et explicabo impedit delectus mollitia.",
+    },
+    {
+        id:12,
+        image: "https://m.media-amazon.com/images/I/71es0+xr4NL._AC_UL480_FMwebp_QL65_.jpg",
+        name: "Stylish Cups",
+        category: "Cups",
+        price: "24.99",
+        description:"Lorem ipsum dolor sit amet consectetur adipisicing elit. Obcaecati, enim aliquid laboriosam eius autem odio minima recusandae fugiat totam aspernatur veritatis laudantium illo alias optio et explicabo impedit delectus mollitia.",
+    },
+
+]
+export default allData;
