+++ src/Components/ItemsList.jsx
@@ -3,15 +3,15 @@ import Items from './Items';
 function ItemsList({ products }) {
     return (
         <div className='flex flex-wrap pl-30 gap-2' >
-            {products && products.map(function (item,idx) {
+            {products && products.map(function (item) {
                 return (
                     <Items
-                        key={idx}
+                        key={item.id}
+                        id={item.id}
                         image={item.image}
                         name={item.name}
                         category={item.category}
                         price={item.price}
-                        showDetails={item.showDetails}
                     />
                 );
             })}
